airflow/requirements.txt
apache-airflow==2.9.0
minio==7.2.3
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.32.3
numpy==1.26.4
openmetadata-ingestion[airflow]==1.5.0